Retain Javadoc for all classes by default.
To limit retention to certain packages, specify a package filter using the
"javadoc.packages" annotation processor option. The value is a comma-delimited
list of package names. When this option is specified, classes outside these
packages (and subpackages, recursively) will be ignored unless they are
annotated with @RetainJavadoc.

Annotation processor options are specified on the javac command line with an
argument of the form:

    -Akey[=value]

For example:

    javac -Ajavadoc.packages=com.example,org.example ...