Remove Elements from method parameters for clarity
These Elements utils are actually used only twice. The rest of the
methods just pass them down to lower level methods.

As Elements are part of the environment, rather than the actual input
of the methods, I find it clearer to avoid passing them as arguments.

This, in turn, makes the use of method references possible in stream
pipelines.