IllegalArgumentException when getting Javadoc for nested class that has no Javadoc #69 A subclass with no Javadoc does not inherit method docs from superclass #70
Loosen the class name check, so it doesn't complain about canonical versus
non-canonical class name. Make it an assertion, so we only pay the cost
when running tests.

When scanning for inherited Javadoc, rework the "isEmpty" check.
Instead of bailing out early if the subclass has no Javadoc,
Wait until after the scan is complete.